+++ src/main/java/gov/ca/cwds/rest/services/ContactIntakeApiService.java
@@ -5,14 +5,14 @@ import gov.ca.cwds.data.ns.ContactDao;
 import gov.ca.cwds.data.ns.ParticipantDao;
 import gov.ca.cwds.data.persistence.ns.ContactEntity;
 import gov.ca.cwds.rest.api.Response;
+import gov.ca.cwds.rest.api.domain.investigation.contact.PostedContactIntake;
 import gov.ca.cwds.rest.api.domain.investigation.contact.ContactIntake;
-import gov.ca.cwds.rest.api.domain.investigation.contact.ContactIntakeRequest;
 import gov.ca.cwds.rest.services.mapper.ContactIntakeMapper;
 import org.apache.commons.lang3.NotImplementedException;
 
 import java.util.Set;
 
-public class ContactIntakeApiService implements TypedCrudsService<String, ContactIntakeRequest, Response>{
+public class ContactIntakeApiService implements TypedCrudsService<String, ContactIntake, Response>{
 
   @Inject
   private ContactDao contactDao;
@@ -27,9 +27,9 @@ public class ContactIntakeApiService implements TypedCrudsService<String, Contac
   }
 
   @Override
-  public Response create(ContactIntakeRequest request) {
+  public Response create(ContactIntake request) {
     ContactEntity entity = ContactIntakeMapper.INSTANCE.map(request);
-    ContactIntake response = ContactIntakeMapper.INSTANCE.map(contactDao.create(entity));
+    PostedContactIntake response = ContactIntakeMapper.INSTANCE.map(contactDao.create(entity));
     Set<String> participantIds = participantDao.findLegacyIdListByScreeningId(request.getScreeningId());
     response.setPeopleIds(participantIds);
     return response;
@@ -50,7 +50,7 @@ public class ContactIntakeApiService implements TypedCrudsService<String, Contac
   }
 
   @Override
-  public Response update(String s, ContactIntakeRequest contactIntake) {
+  public Response update(String s, ContactIntake contactIntake) {
     throw new NotImplementedException("Update is not implemented");
   }
 }
