+++ src/main/java/gov/ca/cwds/rest/services/mapper/ContactIntakeMapper.java
@@ -1,8 +1,8 @@
 package gov.ca.cwds.rest.services.mapper;
 
 import gov.ca.cwds.data.persistence.ns.ContactEntity;
+import gov.ca.cwds.rest.api.domain.investigation.contact.PostedContactIntake;
 import gov.ca.cwds.rest.api.domain.investigation.contact.ContactIntake;
-import gov.ca.cwds.rest.api.domain.investigation.contact.ContactIntakeRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 
@@ -12,7 +12,7 @@ public interface ContactIntakeMapper {
 
   ContactIntakeMapper INSTANCE = Mappers.getMapper(ContactIntakeMapper.class);
 
-  ContactIntake map(ContactEntity contactEntity);
+  PostedContactIntake map(ContactEntity contactEntity);
 
-  ContactEntity map(ContactIntakeRequest contactIntakeRequest);
+  ContactEntity map(ContactIntake contactIntakeRequest);
 }
